+++ common/src/main/java/net/bettercombat/api/WeaponAttributesHelper.java
@@ -85,57 +85,6 @@ public class WeaponAttributesHelper {
         }
     }
 
-    public static final String nbtTag = "weapon_attributes";
-    public static WeaponAttributes readFromNBT(ItemStack itemStack) {
-        var component = itemStack.get(DataComponentTypes.CUSTOM_DATA);
-        if (component == null) {
-            return null;
-        }
-        var nbt = component.getNbt();
-        var attributedItemStack = (ItemStackNBTWeaponAttributes) ((Object)itemStack);
-        var string = nbt.getString(nbtTag);
-        if (string != null && !string.isEmpty() && !attributedItemStack.hasInvalidAttributes()) {
-            var cachedAttributes = attributedItemStack.getWeaponAttributes();
-            if(cachedAttributes != null) {
-                // System.out.println("NBT Attributes - Cache");
-                return cachedAttributes;
-            }
-            Identifier itemId = Registries.ITEM.getId(itemStack.getItem());
-            try {
-                var json = new StringReader(string);
-                var container = decode(json);
-                var attributes = WeaponRegistry.resolveAttributes(itemId, container);
-                if (attributes == null) {
-                    attributedItemStack.setInvalidAttributes(true);
-                }
-                attributedItemStack.setWeaponAttributes(attributes);
-                // System.out.println("NBT Attributes - Resolved");
-                return attributes;
-            } catch (Exception e) {
-                System.err.println("Failed to resolve weapon attributes from ItemStack of item: " + itemId);
-                System.err.println(e.getMessage());
-                attributedItemStack.setInvalidAttributes(true);
-            }
-        }
-        return null;
-    }
-
-    public static void writeToNBT(ItemStack itemStack, AttributesContainer container) {
-        Identifier itemId = Registries.ITEM.getId(itemStack.getItem());
-        var attributedItemStack = (ItemStackNBTWeaponAttributes) ((Object)itemStack);
-        var component = itemStack.getOrDefault(DataComponentTypes.CUSTOM_DATA, NbtComponent.DEFAULT);
-        try {
-            var nbt = component.getNbt();
-            var json = encode(container);
-            nbt.putString(nbtTag, json);
-            attributedItemStack.setInvalidAttributes(false);
-            attributedItemStack.setWeaponAttributes(null);
-        } catch (Exception e) {
-            System.err.println("Failed to write weapon attributes to ItemStack of item: " + itemId);
-            System.err.println(e.getMessage());
-        }
-    }
-
     private static Type attributesContainerFileFormat = new TypeToken<AttributesContainer>() {}.getType();
 
     public static AttributesContainer decode(Reader reader) {